Refactor tests for atomics and generics to use snapshots
- Create `src/tests/semantic_atomic.rs` for C11 atomic builtin tests.
- Create `src/tests/semantic_generic.rs` for C11 `_Generic` selection tests.
- Migrate existing tests from `src/tests/semantic_builtins.rs` to new modules.
- Convert migrated tests to use `cargo-insta` snapshots for MIR and diagnostics verification.
- Update `src/tests.rs` to register new modules.
